Вынесла в отдельный файл Class TextFieldWithPadding, задание решено!
