+++ Navigation/TextFieldWithPadding.swift
@@ -1,0 +1,28 @@
+//
+//  TextFieldWithPadding.swift
+//  Navigation
+//
+//  Created by Irina Bashkova on 30.11.2023.
+//
+
+import UIKit
+
+class TextFieldWithPadding: UITextField {
+    var textPadding = UIEdgeInsets(
+        top: 10,
+        left: 10,
+        bottom: 10,
+        right: 10
+    )
+    
+    override func textRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = super.textRect(forBounds: bounds)
+        return rect.inset(by: textPadding)
+    }
+    
+    override func editingRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = super.editingRect(forBounds: bounds)
+        return rect.inset(by: textPadding)
+    }
+}
+
